Amendment 25 2072-04-20
Objectives

Restrict property rights and compensation in case the state takes over private property. However, the Supreme Court quashed a part of Article 31C, to the extent it took away the power of judicial review. This was done in the landmark case of Kesavananda Bharati v. State of Kerala (1973) 4 SCC 225 which for the first time enunciated the Basic structure doctrine.

Amendments

Amend article 31.
Insert article 31C.[29]

Closes #29

Signed-off-by: sushmit1 <sushmit718@gmail.com>